+++ core/components/romanescobackyard/docs/changelog.md
@@ -3,6 +3,7 @@
 ## Backyard 1.0.0-beta4
 Released on ...
 
+- Prevent critical CSS generator from messing up data urls
 - Add public function to get cache busting string
 - Fix variables in Collection renderers
 - Fix incorrect reference to permissions key in menu config